CORDA-3848 Complete future in flow exception handler
When an exception propagates all the way through the flow to the flow
default uncaught exception handler, the flow should have it's future
completed with an exception.

The `liveFibers` should be decremented as well, since this flow has
technically finished since it exploded.

`FlowExternalAsyncOperationTest` was experiencing a quasar issue when
trying to serialize an external operation. Making the exception handler
change allows this behaviour to be asserted.